requirements.txt
django-icons-tango==1.0